requirements/windows.txt
spyder>=6.1.0,<7.0.0
pywinpty
tornado
coloredlogs
requests
terminado>=0.13.1
pyqt>=5.12,<6
